ДЗ 12-го урока. Произведена замена int-овых цен на BigDecimal , к заказам добавлен статус оплачен/не оплачен. В связи с тем, что при попытке зарегистрировать на PayPal свой рублевый аккаунт, возникли сложности, воспользовалась дефолтным созданным системой аккаунтом, поэтому магазин внезапно стал валютным. На текущем этапе (в рамках учебного проекта) думаю, не критично, но уже отдельно буду с этим разбираться.

+++ api/src/main/java/ru/gb/pugacheva/webapp/api/dtos/CartDto.java
@@ -1,10 +1,11 @@
 package ru.gb.pugacheva.webapp.api.dtos;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class CartDto {
     private List<OrderItemDto> items;
-    private int totalPrice;
+    private BigDecimal totalPrice;
 
     public List<OrderItemDto> getItems() {
 
@@ -15,18 +16,18 @@ public class CartDto {
         this.items = items;
     }
 
-    public int getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
 
     public CartDto() {
     }
 
-    public CartDto(List<OrderItemDto> items, int totalPrice) {
+    public CartDto(List<OrderItemDto> items, BigDecimal totalPrice) {
         this.items = items;
         this.totalPrice = totalPrice;
     }
